model/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
tqdm==4.67.3
scikit-learn==1.4.2
xgboost==3.2.0
matplotlib==3.10.8